pyproject.toml
[tool.poetry]
name = "python-ai-chatbot"
version = "0.1.0"
description = ""
authors = ["Kazuhiro Obama <obamakazuhiro@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
python-dotenv = "^1.0.1"
openai = "^1.59.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
